Use aiokafka's ConsumerRecord type to match its API
This continues using our own Message type internally, but converts
to ConsumerRecord in the consumer so that user code using our
fake consumer lines up with the library we're faking.

As part of this we introduce a default timestamp so that that field
is never empty. The type signature already seemed to expect this,
however actually allowed None values.

Fixes https://github.com/alm0ra/mockafka-py/issues/109.